refactor: move randomString to utils
