+++ lib/workbox/options.js
@@ -1,5 +1,5 @@
 const path = require('path')
-const { joinUrl, getRouteParams, startCase } = require('../utils')
+const { joinUrl, getRouteParams, startCase, randomString } = require('../utils')
 const defaults = require('./defaults')
 
 function getOptions (nuxt, pwa) {
@@ -125,18 +125,6 @@ function getOptions (nuxt, pwa) {
   }
 
   return options
-}
-
-function randomString (length) {
-  const result = []
-  const characters = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789'
-
-  for (let i = 0; i < length; i++) {
-    const char = characters.charAt(Math.floor(Math.random() * characters.length))
-    result.push(char)
-  }
-
-  return result.join('')
 }
 
 module.exports = {
