+++ lib/workbox/utils.js
@@ -29,8 +29,21 @@ async function copyTemplate ({ src, dst, options }) {
   await writeFile(dst, compile({ options }))
 }
 
+function randomString (length) {
+  const result = []
+  const characters = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789'
+
+  for (let i = 0; i < length; i++) {
+    const char = characters.charAt(Math.floor(Math.random() * characters.length))
+    result.push(char)
+  }
+
+  return result.join('')
+}
+
 module.exports = {
   readJSFiles,
   pick,
-  copyTemplate
+  copyTemplate,
+  randomString
 }
